Updated scrapper to use ts-node
