+++ lib/bookstore-scrapper/prisma-client/prismaClient.ts
@@ -1,4 +1,4 @@
 // Instantiate new prisma client
-import pkg from '@prisma/client';
-const { PrismaClient } = pkg;
+import { PrismaClient } from "@prisma/client";
+
 export const prisma = new PrismaClient();
