+++ package.json
@@ -20,7 +20,9 @@
     "cy:run": "dotenv -c test cypress run",
     "cy:dev": "dotenv -c test cypress open",
     "docker:up": "docker-compose up -d",
-    "docker:down": "docker-compose down -v"
+    "docker:down": "docker-compose down -v",
+    "index": "ts-node ./lib/bookstore-scrapper/indexTextbooks.ts",
+    "index:dev": "dotenv -c development run-s index"
   },
   "dependencies": {
     "@prisma/client": "^3.5.0",
@@ -46,6 +48,7 @@
     "@testing-library/react": "^12.1.2",
     "@types/cookie": "^0.4.1",
     "@types/jest": "^27.0.3",
+    "@types/lodash": "^4.14.178",
     "@types/multer": "^1.4.7",
     "@types/node": "^16.11.7",
     "@types/react": "^17.0.35",
